refactor: Schedule components in own files
