+++ src/pages/t/schedule.tsx
@@ -22,7 +22,7 @@ import {
 } from "@suankularb-components/react";
 
 // Components
-import Schedule from "@components/Schedule";
+import Schedule from "@components/schedule/Schedule";
 import DiscardDraft from "@components/dialogs/DiscardDraft";
 
 // Types
